+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improved transaction form.
 - Set 'save' buttons in action bar to 'disabled' state initially.
 - Removed extra 'save' buttons from forms.
+- Moved 'scroll to bottom' button to action bar, added 'scroll to top' button.
 
 ## 0.4.0